requirements.txt
django==3.1
djangorestframework==3.11.1
drf-nested-routers==0.91
django-cors-headers==3.4.0
coverage==5.2.1
gunicorn==20.0.4
factory_boy
unittest2
unittest-xml-reporting
psycopg2-binary